+++ scripts/build-and-deploy.sh
@@ -9,7 +9,7 @@ echo "Deploying SOLR configuration"
 scp -r ~/coding/clojure/priceous/resources/solr/whisky root@$SERVER_IP:/root/solr-6.4.1/server/solr/
 
 echo "Deploying jar"
-scp ~/coding/clojure/priceous/target/priceous-0.2.0-standalone.jar root@$SERVER_IP:/root/priceous/priceous-0.2.0.jar
+scp ~/coding/clojure/priceous/target/priceous-0.3.0-standalone.jar root@$SERVER_IP:/root/priceous/priceous-0.3.0.jar
 
 echo "Copying configuration"
 scp ~/coding/clojure/priceous/resources/priceous.edn root@$SERVER_IP:/root/priceous/