+++ scripts/gather.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+java -cp target/priceous-0.3.0-standalone.jar priceous.core "$@"