+++ scripts/start.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-nohup java -jar priceous-0.2.0.jar ./priceous.edn &
+nohup java -jar priceous-0.3.0.jar ./priceous.edn &